+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Set cache of shield.io badges to 1h
 - Use master/devel branch instead of master
 - Cleanup CONTRIBUTING.md
+- Make the reference implementation use the base token
 
 ### Added
 - [README]: EIP badge with link to the official EIP
@@ -22,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Coverage badge
 - Contract deployment in test using web3.js 1.0.0
 - More tests
+- Add Base Token (pure ERC77 and ERC20 compat.) to let others easily implement their token.
 
 ### Changed
 - Fix typos in [README]